feat(EMS-607): revert moving prepare application tasks function into a directory
